Modified scripts for updated model setups, to allow 'rte' and 'wf_lzs' resume files and updated input files (checks to see if files exist, so should still be compatible with the old setups). Added the '-c' option in the grabber scripts to skip downloading data downloads already completed in the 'GRIB' folder. Changed get_gem_forecast.sh to expect date, instead of using 'yesterday' function as before.
git-svn-id: svn://watflood.uwaterloo.ca/mesh/Branches/people/Dominique/scripts@1388 04c54e49-724f-4ac9-b36f-5436cbaa49b4

+++ ec2_Yukon_GDPS/setup_capa_hindcast_mod.sh
@@ -53,11 +53,12 @@ do
 
    # Obtain other mesh files that are unchanged from the previous gem forecast run
    cp ${input_file_path_gem}/MESH_drainage_database.r2c ${output_file_path_capa}
-   cp ${input_file_path_gem}/MESH_input_reservoir.txt ${output_file_path_capa}
+   [[ -f ${input_file_path_gem}/MESH_input_reservoir.txt ]] && cp ${input_file_path_gem}/MESH_input_reservoir.txt ${output_file_path_capa}
    cp ${input_file_path_gem}/MESH_input_soil_levels.txt ${output_file_path_capa}
    cp ${input_file_path_gem}/MESH_parameters_CLASS.ini ${output_file_path_capa}
    cp ${input_file_path_gem}/MESH_parameters_hydrology.ini ${output_file_path_capa}
-   cp ${input_file_path_gem}/minmax_parameters.txt ${output_file_path_capa}
+   [[ -f ${input_file_path_gem}/MESH_parameters.r2c ]] && cp ${input_file_path_gem}/MESH_parameters.r2c ${output_file_path_capa}
+   [[ -f ${input_file_path_gem}/minmax_parameters.txt ]] && cp ${input_file_path_gem}/minmax_parameters.txt ${output_file_path_capa}
    cp ${input_file_path_gem}/sa_mesh ${output_file_path_capa}
    cp ${input_file_path_gem}/MESH_input_streamflow.txt ${output_file_path_capa} # streamflow values need changing
 
@@ -66,7 +67,13 @@ do
    cp ${input_file_path_capa}/MESH_input_run_options.ini ${output_file_path_capa}
 
    # Obtain mesh state files from the previous capa run
-   if [ -f ${input_file_path_capa}/int_statVariables.seq.wf_route ]; then
+   if [ -f ${input_file_path_capa}/int_statVariables.seq.rte ]; then
+      for name in rte runclass36 lzsp.wfqlz
+	  do
+         cp ${input_file_path_capa}/int_statVariables.seq.${name}_$(date -d "$dt - 1 day" -u +%Y%m%d)16 ${output_file_path_capa}
+         cp ${output_file_path_capa}/int_statVariables.seq.${name}_$(date -d "$dt - 1 day" -u +%Y%m%d)16 ${output_file_path_capa}/int_statVariables.seq.${name}
+	  done
+   elif [ -f ${input_file_path_capa}/int_statVariables.seq.wf_route ]; then
       cp ${input_file_path_capa}/int_statVariables.seq.wf_route_$(date -d "$dt - 1 day" -u +%Y%m%d)16 ${output_file_path_capa}
       cp ${output_file_path_capa}/int_statVariables.seq.wf_route_$(date -d "$dt - 1 day" -u +%Y%m%d)16 ${output_file_path_capa}/int_statVariables.seq.wf_route
       cp ${input_file_path_capa}/int_statVariables.seq.runclass36_$(date -d "$dt - 1 day" -u +%Y%m%d)16 ${output_file_path_capa}
